Rename calculateWeight.js to calculate-weight.js
Switch the final camelCase outlier under src/ to kebab-case so the
naming convention is consistent across the whole library. The
exported function name (calculateWeight) and the public option key
of the same name are unchanged — this is a file-rename only.
